Update discord.gd
I ran into an issue where the data object was blank. This lead to my app throwing an error saying the JSON file had an issue on line 0. I added a print statement on the above method trimming the data object to only a hundred characters and then noticed that when I ran my app a blank line would print out which is when this error would display.

To fix this I've simply added a check and if the string is empty returned null. I had some issues with the editor not liking it when I just returned a normal null, so I moved it to a variable. This can probably be cleaned up and made to look nicer, but I simply wanted to let you know about the problem.